+++ packages/gallery/src/components/item/videos/videoItem.js
@@ -3,7 +3,6 @@ import utils from '../../../common/utils';
 import window from '../../../common/window/windowWrapper';
 import { GalleryComponent } from '../../galleryComponent';
 import EVENTS from '../../../common/constants/events';
-import LAZY_LOAD from '../../../common/constants/lazyLoad';
 import { URL_TYPES, URL_SIZES } from '../../../common/constants/urlTypes';
 import PlayBackground from '../../svgs/components/play_background';
 import PlayTriangle from '../../svgs/components/play_triangle';
@@ -125,7 +124,6 @@ class VideoItem extends GalleryComponent {
     const videoDimensionsCss = {
       width: isWiderThenContainer ? '100%' : 'auto',
       height: isWiderThenContainer ? 'auto' : '100%',
-      opacity: this.props.loadingStatus ? '1' : '0',
     };
 
     if (
@@ -217,44 +215,6 @@ class VideoItem extends GalleryComponent {
     }
   }
 
-  createImageElement() {
-    const {
-      alt,
-      loadingStatus,
-      imageDimensions,
-      createUrl,
-      id,
-      lazyLoad,
-      styleParams,
-    } = this.props;
-
-    return (
-      lazyLoad === LAZY_LOAD.NATIVE ? <img
-        key={
-          (styleParams.cubeImages && styleParams.cubeType === 'fill'
-            ? 'cubed-'
-            : '') + 'image'
-        }
-        className={'gallery-item-visible gallery-item gallery-item-preloaded'}
-        alt={alt ? alt : 'untitled image'}
-        src={createUrl(URL_SIZES.RESIZED, URL_TYPES.HIGH_RES)}
-        loading="lazy"
-        style={imageDimensions}
-      />
-      :
-      <canvas
-        key={'image-' + id}
-        alt={alt ? alt : 'untitled video'}
-        className={
-          'gallery-item-hidden gallery-item-visible gallery-item ' +
-          (loadingStatus.loaded ? ' gallery-item-loaded ' : '') +
-          (loadingStatus.failed ? ' failed ' : '')
-        }
-        data-src={createUrl(URL_SIZES.RESIZED, URL_TYPES.HIGH_RES)}
-      />
-    );
-  }
-
   canVideoPlayInGallery(itemClick, videoPlay , hasLink) {
       if (
         videoPlay === 'hover' || videoPlay === 'auto'
@@ -272,8 +232,7 @@ class VideoItem extends GalleryComponent {
   render() {
     let baseClassName =
       'gallery-item-content gallery-item-visible gallery-item-preloaded gallery-item-video gallery-item video-item' +
-      (utils.isiPhone() ? ' ios' : '') +
-      (this.props.loadingStatus.loaded ? ' gallery-item-loaded ' : '');
+      (utils.isiPhone() ? ' ios' : '');
     if (this.state.playing) {
       baseClassName += ' playing';
     }