+++ packages/gallery/src/settings/merged.js
@@ -4,7 +4,9 @@ import isRelevant from './old/isRelevant';
 import newGallerySettings from './index';
 import defaultStyles from '../common/defaultStyles';
 
-const styleList = [...Array.from(new Set([...Object.keys(newGallerySettings), ...Object.keys(isRelevant), ...Object.keys(content)]))];
+const stylesSet = new Set([...Object.keys(newGallerySettings), ...Object.keys(isRelevant), ...Object.keys(content)]);
+const styleList = [];
+stylesSet.forEach(style => styleList.push(style))
 const mergedSettings = styleList.reduce((obj, styleParam) => {
     const settingsData = newGallerySettings[styleParam] || {
       styleParam,