+++ packages/gallery/src/versionLogger.js
@@ -1,1 +1,1 @@
-try {console.debug('Pro Gallery Version: 1.10.17'); } catch (e) {}
+try {console.debug('Pro Gallery Version: 2.0.2'); } catch (e) {}