+++ packages/layouts/src/group.js
@@ -117,6 +117,7 @@ export class Group {
     const shouldUseFixedHeight =
       this.cubeImages &&
       this.groupSize === 1 &&
+      ['fill', 'fit'].includes(this.cubeType) &&
       this.rotatingGroupTypes.length === 0 &&
       this.rotatingCropRatios.length === 0;
     this.cubedHeight = shouldUseFixedHeight ? height : null;
@@ -745,38 +746,6 @@ export class Group {
       item.groupOffset = {
         left,
         right: left + this.width,
-      };
-    }
-  }
-
-  calcVisibilities(bounds) {
-    if (bounds === true || this.showAllItems === true) {
-      this.onscreen = this.visible = this.rendered = this.required = true;
-    } else if (this.oneRow) {
-      this.onscreen =
-        this.right >= bounds.onscreenTop && this.left <= bounds.onscreenBottom;
-      this.visible =
-        this.right >= bounds.visibleTop && this.left <= bounds.visibleBottom;
-      this.rendered =
-        this.right >= bounds.renderedTop && this.left <= bounds.renderedBottom;
-      this.required =
-        this.right >= bounds.requiredTop && this.left <= bounds.requiredBottom;
-    } else {
-      this.onscreen =
-        this.bottom >= bounds.onscreenTop && this.top <= bounds.onscreenBottom;
-      this.visible =
-        this.bottom >= bounds.visibleTop && this.top <= bounds.visibleBottom;
-      this.rendered =
-        this.bottom >= bounds.renderedTop && this.top <= bounds.renderedBottom;
-      this.required =
-        this.bottom >= bounds.requiredTop && this.top <= bounds.requiredBottom;
-    }
-    for (let i = 0; i < this.items.length; i++) {
-      this.items[i].visibility = {
-        onscreen: this.onscreen,
-        visible: this.visible,
-        rendered: this.rendered,
-        required: this.required,
       };
     }
   }